Small refactor to ErrorLogger in zimcheck
addError() has now a better name: addReportMsg()
The overal status of the test now only depends on "error tests". In other words, tests that produce warning messages (such as Redundant test) are not considered for determining the overalStatus of the tests.
The report() function also do not need to depend on the status of the test. If it has report messages it should print them.